Disable plots with borked data dependencies
The plots for cumulative deaths, daily incidence (of disease), and cumulative
incidence (of disease) are now disabled (fix #4). While this does not rewrite
the plots to not depend on the borked data dependency, there is little else to
do without reengineering a larger portion of the codebase at this time. For now,
the best fix is to step the feature back because it was rushed and wasn't ready
to begin with.