Refactor the UI; write applicative code for seed data upload
The UI is now written as multiple objects, with the ui object created last per
convention and the necessity of the R reader logic.

The changes in server include a single renderUI call applying a function to
generate the seed data upload input widgets. The logic for the input widgets is
not yet implemented.