Display transmission path ui only for Democratic Republic of Congo
Ashok requested that the transmission path file upload inputs only display for
the DRC, so I have made that change. The two file inputs have been moved from
the server to the UI and wrapped in a conditionalPanel that checks that the
selected country is "Democratic Republic of Congo", which works (fix #22).