Clean global.R of dirty and spaghetti code
`source` is removed; it should never be used in production application code,
especially when Shiny already loads everything in the R/ subfolder.

`suppressPackageStartupMessages` is misused; this is corrected to use an
expression so that it is called once.

_here_ is used to setup the use of `here` in the rest of the application.

New commentary on various dirty codes are included.

Code, dirty or not, is restyled for readability and maintainability.

A call to `shinyAppDir` is included to ensure the application starts when the
global.R file is sourced in various editors (VS Code, Positron, Emacs, VIM, etc)
other than RStudio.