Fix 'no such country' error
As mentioned by @bryce-carson in #25 ne_states function was receiving an abbreviation of the name of country this was because I was retrieving the name of the country from the ne_countries 'country' and not countrycode. The first choice would be to use the already existing iso3c code but ne_states only accepts iso_a2 (same as iso2c) so I just used contrycode() to convert between the two.