Reorganize UI code and correct dataAssimilation_* reads
This commit reorganizes the UI for the simulation side of things to reduce some
of the disorganization, but it could still use more organization once things are
done on the integration side of things in server.R.

The integration in server.R is proceeding smoothly; the latest error is
"Warning: Error in <Anonymous>: 2018-10-01 is not prior to or equal to
2018-08-05.", which is a sign that the application has proceeded well into the
beginning of the SVEIRD w/ BDA simulation. I can close the open issue now (fix #50).

The `seedData` identifier is renamed to `uploadedSeedData` throughout some of
the plotting functions to reflect what should be passed to these functions,
given the renamed identifier (`uploadedSeedData` in the UI and server).